+++ DigitalLearningSolutions.Web.Tests/TestHelpers/DiagnosticAssessmentHelper.cs
@@ -31,7 +31,7 @@
             string applicationName = "application name",
             string customisationName = "customisation name",
             string sectionName = "section name",
-            string diagAssessPath = "https://www.dls.nhs.uk/tracking/MOST/Word10Core/Assess/L2_Word_2010_Diag_1.dcr",
+            string diagAssessPath = "https://www.dls.nhs.uk/CMS/CMSContent/Course119/Diagnostic/07DIAGNEW/itspplayer.html",
             bool diagObjSelect = true,
             int plaPassThreshold = 50,
             int currentVersion = 1
