+++ DigitalLearningSolutions.Web/Helpers/ContentViewerHelper.cs
@@ -1,0 +1,10 @@
+﻿namespace DigitalLearningSolutions.Web.Helpers
+{
+    using System.Text.RegularExpressions;
+
+    public static class ContentViewerHelper
+    {
+        public static readonly Regex ScormRegex = new Regex(@".*imsmanifest\.xml$");
+        public static bool IsScormPath(string path) => ScormRegex.IsMatch(path);
+    }
+}
