+++ DigitalLearningSolutions.Web/Views/LearningMenu/Shared/_JavaScriptDisabledError.cshtml
@@ -1,0 +1,11 @@
+﻿@model string
+<div class="nhsuk-error-summary non-js-only" aria-labelledby="error-summary-title" role="alert" tabindex="-1">
+  <h2 class="nhsuk-error-summary__title" id="error-summary-title">
+    There is a problem
+  </h2>
+  <div class="nhsuk-error-summary__body">
+    <p class="nhsuk-u-margin-0">
+      JavaScript is required to view this @Model.
+    </p>
+  </div>
+</div>
