+++ DigitalLearningSolutions.Web/Views/LearningMenu/Tutorial/ContentViewer.cshtml
@@ -10,6 +10,7 @@
   ViewData["CustomisationId"] = Model.CustomisationId;
   ViewData["HeaderPath"] = $"{Configuration["AppRootPath"]}/LearningMenu/{Model.CustomisationId}";
   ViewData["HeaderPathName"] = Model.CourseTitle;
+  const string contentType = "tutorial";
 }
 @section NavMenuItems {
   <partial name="Shared/_NavMenuItems" />
@@ -38,16 +39,7 @@
   <div class="nhsuk-grid-column-full">
     <h1 id="page-heading" class="nhsuk-heading-xl">@Model.TutorialName</h1>
 
-    <div class="nhsuk-error-summary non-js-only" aria-labelledby="error-summary-title" role="alert" tabindex="-1">
-      <h2 class="nhsuk-error-summary__title" id="error-summary-title">
-        There is a problem
-      </h2>
-      <div class="nhsuk-error-summary__body">
-        <p class="nhsuk-u-margin-0">
-          JavaScript is required to view this tutorial.
-        </p>
-      </div>
-    </div>
+    <partial name="Shared/_JavaScriptDisabledError" model="contentType" />
 
     <iframe src="@Model.ContentSource" class="nhsuk-grid-column-full nhsuk-u-margin-bottom-6 js-only-block" height="800">
     </iframe>
